refactor: remove PendingGame and StartedGame
